style: format with ruff 0.15.16 to match the pre-commit pin
The previous format pass used a locally installed ruff 0.15.2, whose
output differs from the 0.15.16 pinned in .pre-commit-config.yaml,
so the hook kept reformatting on the PR. This applies the pinned
version''s formatting, including notebook code cells, which 0.15.16
formats natively.

+++ hamon/conditional_samplers.py
@@ -193,4 +193,6 @@ class SoftmaxConditional(AbstractParametricConditionalSampler):
         output_sd: PyTree[jax.ShapeDtypeStruct],
     ) -> tuple[_State, None]:
         """Sample from a softmax distribution given the parameter vector $\theta$."""
-        return jax.random.categorical(key, parameters, axis=-1).astype(output_sd.dtype), sampler_state
+        return jax.random.categorical(key, parameters, axis=-1).astype(
+            output_sd.dtype
+        ), sampler_state
